Feita refatoração, sepando a lógica, seguindo o exercicio :
**9. \[13 pontos] Refatorar estrutura da pasta pages/Comunidade para separação entre lógica e apresentação**
